# api-spec.yaml
openapi: 3.0.0
servers: []
info:
  version: 0.4.6
  title: Reference Data MS
  description: ''
security:
  - accessCode:
      - read
      - write
...
paths:
...
  /service/ref-group/{ref_group_code}/ref_entries/:
    get:
      tags:
        - Service
      summary: |
        Retrieve a set of Ref Entries under a Ref Group
      description: ''
      security: []
      parameters:
        - $ref: '#/components/parameters/Filter'
        - $ref: '#/components/parameters/RefGroupCode'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: object
                properties:
                  Page:
                    description: ''
                    type: integer
                  PageSize:
                    description: ''
                    type: integer
                  Total:
                    description: ''
                    type: integer
                  Result:
                    type: array
                    items:
                      $ref: '#/components/schemas/RefEntry'
        '400':
          $ref: '#/components/responses/400'
        '500':
          $ref: '#/components/responses/500'
  /service/ref-group/{ref_group_code}/ref-entry/{ref_entry_code}:
    get:
      tags:
        - Service
      summary: |
        Retrieve one item from a Ref Group using its code
      description: ''
      security: []
      parameters:
        - $ref: '#/components/parameters/RefGroupCode'
        - $ref: '#/components/parameters/RefEntryCode'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RefEntry'
        '400':
          $ref: '#/components/responses/400'
        '500':
          $ref: '#/components/responses/500'
...
  /service/ref-groups/fast-banks/ref-entries/:
    get:
      tags:
        - Service
      summary: |
        Retrieve a set of Ref Entries under a Ref Group `Fast Bank`
      description: ''
      security: []
      parameters:
        - $ref: '#/components/parameters/Filter'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: object
                properties:
                  Filter:
                    description: ''
                    type: string
                  Page:
                    description: ''
                    type: integer
                  PageSize:
                    description: ''
                    type: integer
                  Total:
                    description: ''
                    type: integer
                  Result:
                    type: array
                    items:
                      $ref: '#/components/schemas/RefEntryFast'
        '400':
          $ref: '#/components/responses/400'
        '500':
          $ref: '#/components/responses/500'
components:
  securitySchemes:
    accessCode:
      type: oauth2
      flows:
        authorizationCode:
          authorizationUrl: 'http://example.com/oauth/auth'
          tokenUrl: 'http://example.com/oauth/token'
          scopes:
            write: allows modifying resources
            read: allows reading resources
  parameters:
    Page:
      name: page
      in: query
      description: ''
      required: false
      schema:
        type: integer
    PageSize:
      name: page_size
      in: query
      description: ''
      required: false
      schema:
        type: integer
    RefGroupCode:
      in: path
      name: ref_group_code
      schema:
        type: string
      required: true
      description: ''
    RefEntryCode:
      in: path
      name: ref_entry_code
      schema:
        type: string
      required: true
      description: ''
...
    Filter:
      name: filter
      in: query
      description: ''
      required: false
      schema:
        type: string
  responses:
...
    400:
      description: ''
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ExtraInfo'
    500:
      description: ''
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ExtraInfo'
  schemas:
    ExtraInfo:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
      required:
        - code
        - message
    Permissions:
      description: ''
      type: array
      items:
        type: string
        enum:
          - EDIT
          - REMOVE
...
    RefEntry:
      description: ''
      type: object
      properties:
        ref_group_code:
          description: ''
          type: string
          minLength: 2
          maxLength: 32
        code:
          description: ''
          type: string
          minLength: 2
          maxLength: 32
        desc:
          description: ''
          type: string
          minLength: 0
          maxLength: 1000
        sec_id:
          description: ''
          type: integer
        custom_fields:
          $ref: '#/components/schemas/CustomFields'
        allow:
          $ref: '#/components/schemas/Permissions'
...
    CustomFields:
      description: ''
      type: object
    RefEntryFast:
      description: ''
      type: object
      properties:
        code:
          description: ''
          type: string
          minLength: 2
          maxLength: 32
        data:
          $ref: '#/components/schemas/CustomFieldsFast'
      required:
        - code
        - data
    CustomFieldsFast:
      description: ''
      type: object
      properties:
        bank_name:
          type: string
        bic_code:
          type: string
        extra_payer_account_types:
          type: array
          items:
            type: string
        extra_recipient_account_types:
          type: array
          items:
            type: string